Fix update integration tests
